refactor: consolidate all crypto functions in packages/commons
Move signPayload, verifySignature, getCompressedPublicKeyFromAppIdentity,
isPoktBech32Address, isHexPublicKey, pubkeyToAddress, normalizeIdentityToAddress
to @igniter/commons/crypto. Both apps and middleman-workflows now import
from the shared package. Tests moved to commons (25 passing).

Also: remove dead bootstrap service from both apps, deduplicate constants.
