+++ SESSION-SUMMARY.md
@@ -1,0 +1,55 @@
+# Session Summary — 2026-03-31
+
+## Bug Report
+User (JonatanM / Nodefleet) reported keys table crashing in provider admin after v0.9.0:
+```
+Uncaught TypeError: Cannot read properties of null (reading 'name')
+```
+
+## Root Cause
+Two related issues found:
+
+### 1. Null-unsafe cell renderer in keys table
+`apps/provider/src/app/admin/(internal)/keys/table/columns.tsx`
+- `addressGroup` can be `null` (key not assigned to any group), but the cell renderer accessed `.name` without null-check
+- `filterFn` also crashed on null `addressGroup`
+
+### 2. Stale delivery data on released keys (v0.9.0 bug)
+`apps/provider-workflows/src/activities/index.ts`
+- When SupplierStatus workflow releases a Delivered/MissingStake key back to Available (after 24h with no on-chain supplier), it cleared `deliveredAt` and `addressGroupId` but **not** `ownerAddress`, `deliveredTo` (delegator), or reward fields
+- This left "ghost" ownership data on Available keys
+
+## Changes Made
+
+### Fix 1: Null-safe addressGroup rendering
+**File:** `apps/provider/src/app/admin/(internal)/keys/table/columns.tsx`
+- Updated `Key` type: `addressGroup` now `| null`
+- Cell: `addressGroup?.name || '-'`
+- filterFn: early return `false` if `!addressGroup`
+
+### Fix 2: Clean all delivery fields when releasing keys
+**File:** `apps/provider-workflows/src/activities/index.ts`
+- Added `update.deliveredTo = null` and `update.ownerAddress = null` to the 24h release block
+
+### Fix 3: Startup cleanup for existing corrupted data
+**File:** `apps/provider-workflows/src/worker.ts`
+- Added `cleanupStaleAvailableKeys()` that runs on worker startup (after bootstrap)
+- Resets `ownerAddress`, `deliveredTo`, `deliveredAt`, `delegatorRevSharePercentage`, `delegatorRewardsAddress` for any Available key that still has owner/delegator set
+
+## Manual DB Fix (for immediate relief before deploy)
+```sql
+SELECT id, address, state, "ownerAddress", "delegator_identity", "deliveredAt"
+FROM keys
+WHERE state = 'available'
+  AND ("ownerAddress" IS NOT NULL OR "delegator_identity" IS NOT NULL);
+
+UPDATE keys
+SET "ownerAddress" = NULL,
+    "delegator_identity" = NULL,
+    "deliveredAt" = NULL,
+    "address_group_id" = NULL,
+    "delegatorRevSharePercentage" = 0,
+    "delegatorRewardsAddress" = ''
+WHERE state = 'available'
+  AND ("ownerAddress" IS NOT NULL OR "delegator_identity" IS NOT NULL);
+```
