+++ apps/provider-workflows/src/activities/index.ts
@@ -171,6 +171,8 @@ export const providerActivities = (dal: DAL, pocketRpcClient: PocketBlockchain) 
           if (key.deliveredAt && key.deliveredAt.getTime() < Date.now() - 24 * 60 * 60 * 1000) {
             update.state = KeyState.Available
             update.deliveredAt = null
+            update.deliveredTo = null
+            update.ownerAddress = null
             update.addressGroupId = null
             update.delegatorRevSharePercentage = 0
             update.delegatorRewardsAddress = ''
