+++ apps/provider/src/app/api/import-suppliers/submit/route.ts
@@ -3,7 +3,7 @@ import { z } from 'zod'
 import { ensureApplicationIsBootstrapped, validateRequestSignature } from '@/lib/utils/routes'
 import { APIResponse } from '@/lib/models/response'
 import { REQUEST_IDENTITY_HEADER } from '@igniter/commons/constants'
-import { pubkeyToAddress, verifySignature } from '@/lib/crypto'
+import { pubkeyToAddress, verifySignature } from '@igniter/commons/crypto'
 import { verifyAdr36Signature } from '@/lib/adr36'
 import {
   assignSuppliersToDelegate,
