+++ packages/commons/package.json
@@ -38,6 +38,7 @@
   "dependencies": {
     "@cosmjs/crypto": "^0.33.1",
     "@cosmjs/encoding": "^0.33.1",
+    "@cosmjs/proto-signing": "^0.33.1",
     "@cosmjs/stargate": "^0.33.1",
     "@igniter/logger": "workspace:*"
   },
